Created also a SampleBaseApplication
and moved getSampleWallet method to the SampleBaseActivity
