+++ receiverSampleApp/src/main/java/org/kinecosystem/sampleReceiverApp/ReceiverApplication.java
@@ -3,19 +3,14 @@ package org.kinecosystem.sampleReceiverApp;
 
 import android.app.Application;
 
+import org.kinecosystem.baseSampleApp.SampleBaseApplication;
 import org.kinecosystem.baseSampleApp.sampleWallet.SampleWallet;
 
-public class ReceiverApplication extends Application {
+public class ReceiverApplication extends SampleBaseApplication {
     private static final String APP_ID = "recv";
-    private SampleWallet sampleWallet;
 
     @Override
-    public void onCreate() {
-        super.onCreate();
-        sampleWallet = new SampleWallet(this, APP_ID);
-    }
-
-    public SampleWallet getSampleWallet(){
-        return sampleWallet;
+    public String getAppId() {
+        return APP_ID;
     }
 }
