--- receiverSampleApp/src/main/res/layout/activity_main.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ReceiverMainActivity">
-
-    <TextView
-        android:id="@+id/publicAddressView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/init_wallet"
-        android:layout_marginStart="16dp"
-        android:layout_marginLeft="16dp"
-        android:layout_marginTop="16dp"
-        android:visibility="visible"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/balanceView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/init_balance"
-        android:layout_marginStart="16dp"
-        android:layout_marginLeft="16dp"
-        android:layout_marginTop="16dp"
-        android:visibility="visible"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/publicAddressView" />
-
-</android.support.constraint.ConstraintLayout>
