+++ senderSampleApp/src/main/java/org/kinecosystem/sampleSenderApp/SenderApplication.java
@@ -1,22 +1,13 @@
 package org.kinecosystem.sampleSenderApp;
 
 
-import android.app.Application;
+import org.kinecosystem.baseSampleApp.SampleBaseApplication;
 
-import org.kinecosystem.baseSampleApp.sampleWallet.SampleWallet;
-
-
-public class SenderApplication extends Application {
+public class SenderApplication extends SampleBaseApplication {
     private static final String APP_ID = "sndr";
-    private SampleWallet sampleWallet;
 
     @Override
-    public void onCreate() {
-        super.onCreate();
-        sampleWallet = new SampleWallet(this, APP_ID);
-    }
-
-    public SampleWallet getSampleWallet(){
-        return sampleWallet;
+    public String getAppId() {
+        return APP_ID;
     }
 }
