+++ senderSampleApp/src/main/java/org/kinecosystem/sampleSenderApp/SenderMainActivity.java
@@ -1,26 +1,17 @@
 package org.kinecosystem.sampleSenderApp;
 
 import android.os.Bundle;
-import android.os.PersistableBundle;
 import android.support.annotation.Nullable;
 
 import org.kinecosystem.appsdiscovery.sender.discovery.view.customView.AppsDiscoveryAlertDialog;
 import org.kinecosystem.baseSampleApp.SampleBaseActivity;
-import org.kinecosystem.baseSampleApp.sampleWallet.SampleWallet;
 
 public class SenderMainActivity extends SampleBaseActivity {
-
 
     @Override
     protected int getLayoutId() {
         return R.layout.activity_sample;
     }
-
-    @Override
-    protected SampleWallet getSampleWallet() {
-        return ((SenderApplication) getApplicationContext()).getSampleWallet();
-    }
-
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
